timemachine/pyproject.toml
[tool.poetry]
name = "timemachine"
version = "0.1.0"
description = ""
authors = ["moon <black2st11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-pghistory = "^1.5.0"
django-phonenumber-field = "^6.3.0"
psycopg2 = "^2.9.3"
phonenumbers = "^8.12.52"
django-dotenv = "^1.4.2"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
